+++ CHANGELOG.md
@@ -5,9 +5,13 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- Add support for service description. (See: `Service::set_description`)
+
 ### Fixed
 - Fix segmentation fault in `Service` functions, that query service config, by moving buffer 
   allocation on heap.
+
 
 ## [0.3.0] - 2020-06-18
 ### Added